# docs/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is HexLabs' API Documentation.",
    "title": "API Reference",
    "version": "1.0.0"
  },
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Search for a user profile.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string",
            "description": "specifies the number of valid objects to return",
            "required": true
          },
          {
            "name": "offset",
            "in": "query",
            "type": "string",
            "description": "specifies offset in pagination",
            "required": true
          },
          {
            "name": "regex",
            "in": "query",
            "type": "string",
            "description": "specifies if search term uses regex or not",
            "required": true
          },
          {
            "name": "search",
            "in": "query",
            "type": "string",
            "description": "search term",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "offset": {
                  "type": "integer"
                },
                "total": {
                  "type": "integer"
                },
                "count": {
                  "type": "integer"
                },
                "profiles": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Profile"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create a user's profile.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "object",
                  "properties": {
                    "first": {
                      "type": "string"
                    },
                    "middle": {
                      "type": "string"
                    },
                    "last": {
                      "type": "string"
                    }
                  }
                },
                "phoneNumber": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "resume": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": ["users"],
        "summary": "Finds and displays profile by user ID.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "integer",
            "description": "User ID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "allOf": [{ "$ref": "#/definitions/Profile" }, { "$ref": "#/definitions/Permission" }]
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update a user's profile.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "integer",
            "description": "User ID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/users/actions/retrieve": {
      "post": {
        "tags": ["users"],
        "summary": "Gets user profiles by list of userIds",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userIds": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Profile"
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "name": {
          "type": "object",
          "properties": {
            "first": {
              "type": "string"
            },
            "middle": {
              "type": "string"
            },
            "last": {
              "type": "string"
            }
          }
        },
        "phoneNumber": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "resume": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Profile"
      }
    },
    "Permission": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "roles": {
          "type": "object",
          "properties": {
            "member": {
              "type": "boolean"
            },
            "exec": {
              "type": "boolean"
            },
            "admin": {
              "type": "boolean"
            }
          }
        }
      },
      "xml": {
        "name": "Permission"
      }
    },
...
  }
}